engine: Introduce view's main engine
Implement an engine that captures keyboard events and then sends them
down a tree of `View`s for processing. A `View` receiving an `Event` may
choose to process it, and then once the entire tree is done processing
the `Event`, the tree is rendered. As such, each `View` is responsible
for rendering itself on a provided `Pane`.

Support the engine with two `View`s:
- `TextView`:  read-only object that displays text.
- `FrameView`: cosmetic object that surrounds a `View` with a border.

The driver is backed by the ncurses library for both event capturing and
output rendering. This may or may not be extended in the future to
support other backends.