views: Implement a terminal-splitting `View`
Introduce a `SplitView` construct the facilitates displaying one `View`
next to another. There's alway one and only one child `View` in focus
per `SplitView`, to which terminal events are delivered. The `View` in
focus can be changed using a combo-key. As such, a child `View` are
discouraged from processing these events unless it is a `SplitView`.
Following this rule allows `SplitView`s compose recursively and can be
traversed across naturally.